feat: Add Streamlit UI and improve project structure
- Implement new user interface using Streamlit framework
- Reorganize file hierarchy for better maintainability
- Add run.sh script to streamline tool execution (API/normal mode)
- Add megapixel camera support to signature and attack files

The run script simplifies tool usage by providing unified entry point
for both API and standard operation modes.
