+++ Genzai-Tool/go.mod
@@ -1,0 +1,3 @@
+module genzai
+
+go 1.17
