--- go.mod
@@ -1,3 +1,0 @@
-module genzai
-
-go 1.17
